curve refatoring
