+++ src/connectors/index.mjs
@@ -2,4 +2,4 @@ export * from './jumpover.mjs';
 export * from './normal.mjs';
 export * from './rounded.mjs';
 export * from './smooth.mjs';
-
+export * from './curve.mjs';
